refactor(scraper): address CodeRabbit findings — extract SourceCooldownTracker, add try_admit, fix Workday 403 fall-through
- Extract SourceCooldownTracker, SOURCE_COOLDOWN_THRESHOLD, and SOURCE_COOLDOWN
  singleton to scripts/source_cooldown.py; update_jobs.py re-exports them for
  backward compatibility (all existing imports continue to work).

- Add atomic SourceCooldownTracker.try_admit(source) -> bool method: under a
  single lock, returns False if already tripped, increments count and trips if
  threshold reached (returning False), returns True if still below threshold.
  Eliminates the TOCTOU gap between separate record_403 + is_tripped calls.

- Update all 4 fetch paths (Greenhouse, Lever, Workday, Google) to call
  try_admit() instead of separate record_403() + is_tripped(), preventing any
  possibility of double-counting under concurrent load.

- Fix Workday 403 fall-through: introduce handled_403 local flag; skip both the
  generic 'Workday API error' log and the '✓ Found N jobs' success footer when
  a 403 was already handled by the cooldown path.

- Tighten test_403_logs_via_cooldown_path in test_workday_422_logging.py to
  assert that 'Workday API error' and '✓ Found' are NOT in captured output.

- Update test_source_cooldown.py imports to pull SourceCooldownTracker,
  SOURCE_COOLDOWN, SOURCE_COOLDOWN_THRESHOLD directly from source_cooldown;
  update_jobs.py integration imports remain unchanged.

All 246 tests pass. Bandit: no new findings in changed files (pre-existing B104
in update_jobs.py unrelated to this change).
